Move OutputStream and derivatives to a separate header
